+++ web/src/main/java/hueHarmony/web/util/FunctionsUtil.java
@@ -3,7 +3,7 @@ package hueHarmony.web.util;
 import org.springframework.context.annotation.Bean;
 
 public class FunctionsUtil {
-    public static int calculateLoyaltyPoints(Double totalAmountSpent) {
+    public static int calculateLoyaltyPoints(float totalAmountSpent) {
         int pointsPerDollar = 1; // 1 point for every $10 spent
         return (int) (totalAmountSpent / 250) * pointsPerDollar;
     }
